fix data types further

+++ cogs/jose.py
@@ -6,13 +6,13 @@ class Jose:
         self.bot = bot
 
     @commands.command(hidden=True)
-    async def joseprob(self, ctx, jccount: int, amount: int):
+    async def joseprob(self, ctx, jccount: float, amount: float):
         prob = (1 + (jccount / amount)) * 0.42
         await ctx.send(f"Probability: {prob}")
 
 
     @commands.command(hidden=True)
-    async def josemax(self, ctx, jccount: int):
+    async def josemax(self, ctx, jccount: float):
         maxfifty = (jccount / 4.58) * 0.42
         await ctx.send(f"Maximum amount that gives 5 probability: {maxfifty}")
